Release v10.0.0

+++ src/spot/websocket_api/apis/account_api.rs
@@ -258,11 +258,11 @@ pub struct AllOrdersParams {
     /// This field is **optional.
     #[builder(setter(into), default)]
     pub id: Option<String>,
-    /// Cancel order by orderId
+    /// `orderId`or`origClientOrderId`mustbesent
     ///
     /// This field is **optional.
     #[builder(setter(into), default)]
-    pub order_id: Option<i32>,
+    pub order_id: Option<i64>,
     ///
     /// The `start_time` parameter.
     ///
@@ -441,11 +441,11 @@ pub struct MyTradesParams {
     /// This field is **optional.
     #[builder(setter(into), default)]
     pub id: Option<String>,
-    /// Cancel order by orderId
+    /// `orderId`or`origClientOrderId`mustbesent
     ///
     /// This field is **optional.
     #[builder(setter(into), default)]
-    pub order_id: Option<i32>,
+    pub order_id: Option<i64>,
     ///
     /// The `start_time` parameter.
     ///
@@ -658,11 +658,11 @@ pub struct OrderStatusParams {
     /// This field is **optional.
     #[builder(setter(into), default)]
     pub id: Option<String>,
-    /// Cancel order by orderId
+    /// `orderId`or`origClientOrderId`mustbesent
     ///
     /// This field is **optional.
     #[builder(setter(into), default)]
-    pub order_id: Option<i32>,
+    pub order_id: Option<i64>,
     /// `orderId`or`origClientOrderId`mustbesent
     ///
     /// This field is **optional.
